Breaking index.js into different files for each platform
